+++ changelog.md
@@ -1,5 +1,23 @@
 # Changelog
 
 **./test/fox-test-main.js**
-* Console is now cleared upon input prompt and not afterwards.
-	* This means that there is only one call to `clear();` at the start of the file instead of in testing modes.
+* The 'debug' mode now calls "Part F - Controller Models"
+
+---
+
+**./app/test-reference.js**
+* Replaced calls:
+	* 'checkObjectSearchValue' with 'readCurrentObject'
+	* 'getJsonObjectValues' with 'getObjectValues'
+
+---
+
+**./test-parts/part-c-internal_scripts/sub-scripts/common-request.js**
+* Replaced 'commonFunctionsFile' with 'arrayFunctions' when calling:
+	* testPopulated
+	* testAllType
+
+---
+
+**./test-parts/part-h-api_requests/requests/req-storage.js**
+* Added missing 'commonPaths.testObject' requirement.